+++ src/cqfantasy/watchtower/__init__.py
@@ -5,3 +5,4 @@ from .TowerBodyGreebled import TowerBodyGreebled
 from .TowerTop import TowerTop
 from .TowerTopGreebled import TowerTopGreebled
 from .WatchTower import WatchTower
+from .HillBase import HillBase